+++ test/conformance/skipped_tests_test.go
@@ -34,14 +34,12 @@ var skippedTestsForHybrid = []string{
 	// Core profile.
 	tests.HTTPRouteMethodMatching.ShortName,
 	tests.HTTPRouteQueryParamMatching.ShortName,
+	// TODO: https://github.com/Kong/kong-operator/issues/4581
+	tests.HTTPRoutePathMatchOrder.ShortName,
 
 	// Extended profile.
 	tests.HTTPRouteRewriteHost.ShortName,
 	tests.HTTPRouteRewritePath.ShortName,
-
-	// This test is skipped because it proven to be flaky.
-	// TODO: https://github.com/Kong/kong-operator/issues/2793
-	tests.HTTPRouteReferenceGrant.ShortName,
 }
 
 // skippedTestsForConfig returns the list of skipped tests for the given router flavor and gateway type.